pyproject.toml
[project]
name = "sumtech-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.119.0",
    "google-generativeai>=0.8.6",
    "huggingface_hub>=0.27.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
